+++ client/app/account/settings/settings.controller.js
@@ -20,7 +20,6 @@ export default class SettingsController {
   changePassword(form) {
     this.submitted = true;
     if(form.$valid) {
-      console.log('form.$valid')
       this.Auth.changePassword(this.user.oldPassword, this.user.newPassword)
         .then(() => {
           this.message = 'Password successfully changed.';