+++ client/app/sysadmin/sysadmin.component.spec.js
@@ -1,0 +1,17 @@
+'use strict';
+
+describe('Component: SysadminComponent', function() {
+  // load the controller's module
+  beforeEach(module('kupikiHotspotAdminApp.sysadmin'));
+
+  var SysadminComponent;
+
+  // Initialize the controller and a mock scope
+  beforeEach(inject(function($componentController) {
+    SysadminComponent = $componentController('sysadmin', {});
+  }));
+
+  it('should ...', function() {
+    expect(1).toEqual(1);
+  });
+});