+++ server/api/user/user.model.js
@@ -175,11 +175,11 @@ export default function(sequelize, DataTypes) {
 
         if(!callback) {
           // eslint-disable-next-line no-sync
-          return crypto.pbkdf2Sync(password, salt, defaultIterations, defaultKeyLength)
+          return crypto.pbkdf2Sync(password, salt, defaultIterations, defaultKeyLength, 'sha256')
                        .toString('base64');
         }
 
-        return crypto.pbkdf2(password, salt, defaultIterations, defaultKeyLength,
+        return crypto.pbkdf2(password, salt, defaultIterations, defaultKeyLength, 'sha256',
           function(err, key) {
             if(err) {
               callback(err);