+++ server/config/environment/index.js
@@ -34,15 +34,6 @@ var all = {
   // Secret for session, you will want to change this and make it an environment variable
   secrets: {
     session: 'kupiki-hotspot-admin-secret'
-  },
-
-  // MongoDB connection options
-  mongo: {
-    options: {
-      db: {
-        safe: true
-      }
-    }
   }
 };