+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+**v0.2**
+- Improve login screen
+- Add random background
+- Cleaning of coreui default views
+- Add view for system management
+
+**v0.1c**
+- Add support of NodeJS v8.4.0
+- User password encrypted using SHA256 digest
+
 **v0.1b**
 - Cleaned version, tested on Raspberry Pi with SQlite database